+++ src/openforms/fixtures/default_admin_index.json
@@ -155,7 +155,7 @@
                 "cookiegroup"
             ],
             [
-                "custom_field_types",
+                "np_family_members",
                 "familymemberstypeconfig"
             ],
             [
