+++ src/openforms/formio/components/np_family_members/models.py
@@ -16,4 +16,3 @@ class FamilyMembersTypeConfig(SingletonModel):
 
     class Meta:
         verbose_name = _("Family members type configuration")
-        db_table = "custom_field_types_familymemberstypeconfig"
